Address public write review boundary cleanup
