+++ packages/cli/test/import-artifact-routes.test.ts
@@ -15,6 +15,7 @@ import {
   keccak256ContentHash,
   sharedMemoryReadBothFilter,
 } from '@origintrail-official/dkg-core';
+import { reconstructChunkedText } from '../../core/test/helpers/chunked-text.js';
 import { FileStore } from '../src/file-store.js';
 import type { ExtractionStatusRecord } from '../src/extraction-status.js';
 import { handleKnowledgeAssetsRoutes } from '../src/daemon/routes/knowledge-assets.js';
@@ -23,27 +24,6 @@ const DKG = 'http://dkg.io/ontology/';
 const PROV = 'http://www.w3.org/ns/prov#';
 
 type Quad = { subject: string; predicate: string; object: string; graph?: string };
-
-function reconstructChunkedText(quads: readonly Quad[], subject: string): string {
-  const bodySubject = quads.find((quad) =>
-    quad.subject === subject &&
-    quad.predicate === DKG_HAS_TEXT_BODY
-  )?.object;
-  if (!bodySubject) throw new Error(`Missing chunked text body for ${subject}`);
-  return quads
-    .filter((quad) => quad.subject === bodySubject && quad.predicate === DKG_HAS_TEXT_CHUNK)
-    .map((link) => {
-      const chunkQuads = quads.filter((quad) => quad.subject === link.object);
-      const indexTerm = chunkQuads.find((quad) => quad.predicate === DKG_CHUNK_INDEX)?.object;
-      const valueTerm = chunkQuads.find((quad) => quad.predicate === DKG_CHUNK_VALUE)?.object;
-      const index = Number(/^"(\d+)"/.exec(indexTerm ?? '')?.[1] ?? NaN);
-      if (!Number.isInteger(index) || !valueTerm) throw new Error(`Invalid chunk ${link.object}`);
-      return { index, value: JSON.parse(valueTerm) as string };
-    })
-    .sort((a, b) => a.index - b.index)
-    .map((chunk) => chunk.value)
-    .join('');
-}
 
 function sha256Hash(bytes: Buffer): string {
   return `sha256:${createHash('sha256').update(bytes).digest('hex')}`;
