+++ packages/cli/test/import-file-integration.part-01.test.ts
@@ -6,27 +6,7 @@ import {
   DKG_HAS_TEXT_BODY,
   DKG_HAS_TEXT_CHUNK,
 } from '@origintrail-official/dkg-core';
-
-function reconstructChunkedText(quads: readonly CapturedQuad[], subject: string): string {
-  const bodySubject = quads.find((quad) =>
-    quad.subject === subject &&
-    quad.predicate === DKG_HAS_TEXT_BODY
-  )?.object;
-  if (!bodySubject) throw new Error(`Missing chunked text body for ${subject}`);
-  return quads
-    .filter((quad) => quad.subject === bodySubject && quad.predicate === DKG_HAS_TEXT_CHUNK)
-    .map((link) => {
-      const chunkQuads = quads.filter((quad) => quad.subject === link.object);
-      const indexTerm = chunkQuads.find((quad) => quad.predicate === DKG_CHUNK_INDEX)?.object;
-      const valueTerm = chunkQuads.find((quad) => quad.predicate === DKG_CHUNK_VALUE)?.object;
-      const index = Number(/^"(\d+)"/.exec(indexTerm ?? '')?.[1] ?? NaN);
-      if (!Number.isInteger(index) || !valueTerm) throw new Error(`Invalid chunk ${link.object}`);
-      return { index, value: JSON.parse(valueTerm) as string };
-    })
-    .sort((a, b) => a.index - b.index)
-    .map((chunk) => chunk.value)
-    .join('');
-}
+import { reconstructChunkedText } from '../../core/test/helpers/chunked-text.js';
 
 describe('import-file orchestration — happy paths', () => {
